# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Citation": {
        "description": "One retrieved chunk backing a claim in the reply.",
        "properties": {
          "chunk_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Index"
          },
          "excerpt": {
            "default": "",
            "title": "Excerpt",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "source"
        ],
        "title": "Citation",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "A native speech-to-speech coach for building communication confidence. Accessibility tool, not a medical device \u2014 see docs/ETHICS.md.",
    "title": "Speech Confidence Coach",
    "version": "contract"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/corpus": {
      "get": {
        "operationId": "corpus_status_api_corpus_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Corpus Status Api Corpus Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Corpus Status",
        "tags": [
          "corpus"
        ]
      }
    },
    "/api/corpus/ingest": {
      "post": {
        "operationId": "ingest_api_corpus_ingest_post",
        "parameters": [
          {
            "description": "Drop the existing index first. Use after removing documents.",
            "in": "query",
            "name": "reset",
            "required": false,
            "schema": {
              "default": false,
              "description": "Drop the existing index first. Use after removing documents.",
              "title": "Reset",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ingest Api Corpus Ingest Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest",
        "tags": [
          "corpus"
        ]
      }
    },
    "/api/corpus/manifest": {
      "get": {
        "description": "What is indexed, by source.\n\nReconcile this against data/corpus/SOURCES.md before writing the report: a\ndocument indexed but unlogged has no provenance, and one logged but\nunindexed is backing no answer.",
        "operationId": "manifest_api_corpus_manifest_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Manifest Api Corpus Manifest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Manifest",
        "tags": [
          "corpus"
        ]
      }
    },
    "/api/corpus/search": {
      "get": {
        "description": "Query retrieval directly.\n\nThe fastest way to check whether a poor answer came from bad retrieval or\nbad generation, which is otherwise hard to tell apart.",
        "operationId": "search_api_corpus_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "k",
            "required": false,
            "schema": {
              "default": 4,
              "maximum": 20,
              "minimum": 1,
              "title": "K",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Citation"
                  },
                  "title": "Response Search Api Corpus Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search",
        "tags": [
          "corpus"
        ]
      }
    },
...
  }
}